lint wip

+++ cmd/lakectl/cmd/output.go
@@ -94,7 +94,7 @@ func WriteTo(tpl string, ctx interface{}, w io.Writer) {
 			if b < unit {
 				return fmt.Sprintf("%d B", b)
 			}
-			div, exp := int64(unit), 0
+			div, exp := unit, 0
 			for n := b / unit; n >= unit; n /= unit {
 				div *= unit
 				exp++
